enterprise-agentic-rag/requirements.txt
# Core
streamlit==1.35.0
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1

# LLM / Agents
openai==1.30.1
langchain==0.2.3
langchain-openai==0.1.8
langchain-community==0.2.3
langgraph==0.1.5

# Vector DB
qdrant-client==1.9.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.30

# Data / Charts
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
matplotlib==3.9.0

# Monitoring
langsmith==0.1.68

# Utils
python-dotenv==1.0.1
httpx==0.27.0
